[MIPS] Make lw/sw work with larger than int16 offset.
Fixes mjsunit array-constructor.js test.

That test pushed large numbers of objects to stack, and got to
point it could no longer reference the top of stack (sp + 32K)
correctly.

Have fixed lb, lbu, lh, lhu, lw, sb, sh, sw.

+++ src/mips/assembler-mips.h
@@ -264,10 +264,10 @@ class Operand BASE_EMBEDDED {
 class MemOperand : public Operand {
  public:
 
-  explicit MemOperand(Register rn, int16_t offset = 0);
+  explicit MemOperand(Register rn, int32_t offset = 0);
 
  private:
-  int16_t offset_;
+  int32_t offset_;
 
   friend class Assembler;
 };
